test(blackbox): split m2o MAX_BATCH_MUTATION into a sequential file
The MAX_BATCH_MUTATION createMany/updateBatch/updateMany tests are the only
WebSocket users in m2o.test.ts: each opens a WS connection + a GraphQL-WS
subscription against the shared blackbox instance. Running in the 6-fork
parallel pool, the WS handshake starves under peak load and blows its 20s
waitTimeout -> 'WebSocket failed to achieve the OPEN state', reddening a whole
shard at random. The dedicated WebSocket suites are already in the sequential
'after' list for exactly this reason; m2o smuggled WS in while parallel.

Extract the MAX_BATCH_MUTATION block into m2o-max-batch-mutation.test.ts and
add it to the sequential 'after' list, so its WS handshakes run without the
concurrent-load starvation. This also removes the shared-instance concurrency
behind the documented intermittent 200-vs-400 count race.

The sharded seeder (seed-database.test.ts) creates each seed's structure only
on its own test file's shard, so the new file ships m2o-max-batch-mutation.seed.ts
re-exporting seedDBStructure -> the collections exist on its shard. Re-running
seedDBStructure is safe: CreateRelation was already GET-first idempotent, and
CreateField now reads the existing field back on the 'already exists' error
instead of failing (fast path unchanged - only a duplicate re-seed pays an
extra GET). The MAX_BATCH tests create their own items, so no seedDBValues.

m2o.test.ts keeps the rest (WS-free, still parallel) and drops the now-unused
config / ReadItem / createWebSocket* imports. The moved block predates the
85-col rule, so 58 grandfathered long lines are wrapped mechanically (verbatim
semantics, no logic change).

The adjacent retry-unsafe mutation tracker (items.ts) is split to #312.

Refs #277
