+++ tests/blackbox/tests/db/routes/items/m2o-max-batch-mutation.seed.ts
@@ -1,0 +1,1 @@
+export { seedDBStructure } from './m2o.seed';
