+++ tests/blackbox/tests/db/routes/items/m2o.test.ts
@@ -1,7 +1,7 @@
-import config, { getUrl } from '@common/config';
-import { CreateItem, ReadItem } from '@common/functions';
+import { getUrl } from '@common/config';
+import { CreateItem } from '@common/functions';
 import vendors from '@common/get-dbs-to-test';
-import { createWebSocketConn, createWebSocketGql, requestGraphQL } from '@common/transport';
+import { requestGraphQL } from '@common/transport';
 import type { PrimaryKeyType } from '@common/types';
 import { PRIMARY_KEY_TYPES, USER } from '@common/variables';
 import { without } from 'lodash-es';
@@ -1202,740 +1202,6 @@ describe.each(PRIMARY_KEY_TYPES)('/items', (pkType) => {
 								}),
 							).toEqual(expectedDesc);
 						});
-					});
-				});
-			});
-
-			describe('MAX_BATCH_MUTATION Tests', () => {
-				describe('createMany', () => {
-					describe('passes when below limit', () => {
-						it.each(vendors)(
-							'%s',
-							async (vendor) => {
-								// Setup
-								const count = Number(config.envs[vendor]['MAX_BATCH_MUTATION']) / 2;
-								const states: any[] = [];
-								const states2: any[] = [];
-
-								for (let i = 0; i < count; i++) {
-									states.push(createState(pkType));
-									states[i].country_id = createCountry(pkType);
-
-									states2.push(createState(pkType));
-									states2[i].country_id = createCountry(pkType);
-								}
-
-								const ws = createWebSocketConn(getUrl(vendor), {
-									auth: { access_token: USER.ADMIN.TOKEN },
-								});
-
-								await ws.subscribe({ collection: localCollectionCountries, uid: localCollectionCountries });
-								await ws.subscribe({ collection: localCollectionStates, uid: localCollectionStates });
-
-								const wsGql = createWebSocketGql(getUrl(vendor), {
-									auth: { access_token: USER.ADMIN.TOKEN },
-								});
-
-								const subscriptionKeyCountries = await wsGql.subscribe({
-									collection: localCollectionCountries,
-									jsonQuery: {
-										event: true,
-										data: {
-											id: true,
-											name: true,
-										},
-									},
-									uid: localCollectionCountries,
-								});
-
-								const subscriptionKeyStates = await wsGql.subscribe({
-									collection: localCollectionStates,
-									jsonQuery: {
-										event: true,
-										data: {
-											id: true,
-											name: true,
-											country_id: {
-												id: true,
-											},
-										},
-									},
-									uid: localCollectionStates,
-								});
-
-								// Action
-								const response = await request(getUrl(vendor))
-									.post(`/items/${localCollectionStates}`)
-									.send(states)
-									.set('Authorization', `Bearer ${USER.ADMIN.TOKEN}`);
-
-								const wsMessagesCountries = await ws.getMessages(count, { uid: localCollectionCountries });
-								const wsMessagesStates = await ws.getMessages(count, { uid: localCollectionStates });
-								const wsGqlMessagesCountries = await wsGql.getMessages(count, { uid: localCollectionCountries });
-								const wsGqlMessagesStates = await wsGql.getMessages(count, { uid: localCollectionStates });
-
-								const mutationKey = `create_${localCollectionStates}_items`;
-
-								const gqlResponse = await requestGraphQL(getUrl(vendor), false, USER.ADMIN.TOKEN, {
-									mutation: {
-										[mutationKey]: {
-											__args: {
-												data: states2,
-											},
-											id: true,
-										},
-									},
-								});
-
-								const wsMessagesGqlCountries = await ws.getMessages(count, { uid: localCollectionCountries });
-								const wsMessagesGqlStates = await ws.getMessages(count, { uid: localCollectionStates });
-								ws.conn.close();
-								const wsGqlMessagesGqlCountries = await wsGql.getMessages(count, { uid: localCollectionCountries });
-								const wsGqlMessagesGqlStates = await wsGql.getMessages(count, { uid: localCollectionStates });
-								wsGql.client.dispose();
-
-								// Assert
-								expect(response.statusCode).toBe(200);
-								expect(response.body.data.length).toBe(count);
-								expect(gqlResponse.statusCode).toBe(200);
-								expect(gqlResponse.body.data[mutationKey].length).toEqual(count);
-
-								for (const { messagesCountries, messagesStates } of [
-									{ messagesCountries: wsMessagesCountries, messagesStates: wsMessagesStates },
-									{ messagesCountries: wsMessagesGqlCountries, messagesStates: wsMessagesGqlStates },
-								]) {
-									expect(messagesCountries?.length).toBe(count);
-									expect(messagesStates?.length).toBe(count);
-
-									for (let i = 0; i < count; i++) {
-										expect(messagesCountries![i]).toMatchObject({
-											type: 'subscription',
-											event: 'create',
-											data: [
-												{
-													id: expect.anything(),
-													name: expect.any(String),
-												},
-											],
-										});
-
-										expect(messagesStates![i]).toMatchObject({
-											type: 'subscription',
-											event: 'create',
-											data: [
-												{
-													id: expect.anything(),
-													name: expect.any(String),
-													country_id: expect.anything(),
-												},
-											],
-										});
-									}
-								}
-
-								for (const { messagesCountries, messagesStates } of [
-									{ messagesCountries: wsGqlMessagesCountries, messagesStates: wsGqlMessagesStates },
-									{ messagesCountries: wsGqlMessagesGqlCountries, messagesStates: wsGqlMessagesGqlStates },
-								]) {
-									expect(messagesCountries?.length).toBe(count);
-									expect(messagesStates?.length).toBe(count);
-
-									for (let i = 0; i < count; i++) {
-										expect(messagesCountries![i]).toEqual({
-											data: {
-												[subscriptionKeyCountries]: {
-													event: 'create',
-													data: {
-														id: expect.anything(),
-														name: expect.any(String),
-													},
-												},
-											},
-										});
-
-										expect(messagesStates![i]).toEqual({
-											data: {
-												[subscriptionKeyStates]: {
-													event: 'create',
-													data: {
-														id: expect.anything(),
-														name: expect.any(String),
-														country_id: {
-															id: expect.anything(),
-														},
-													},
-												},
-											},
-										});
-									}
-								}
-							},
-							120000,
-						);
-					});
-
-					describe('errors when above limit', () => {
-						it.each(vendors)(
-							'%s',
-							async (vendor) => {
-								// Setup
-								const count = Number(config.envs[vendor]['MAX_BATCH_MUTATION']) / 2 + 1;
-								const states: any[] = [];
-								const states2: any[] = [];
-
-								for (let i = 0; i < count; i++) {
-									states.push(createState(pkType));
-									states[i].country_id = createCountry(pkType);
-
-									states2.push(createState(pkType));
-									states2[i].country_id = createCountry(pkType);
-								}
-
-								const ws = createWebSocketConn(getUrl(vendor), {
-									auth: { access_token: USER.ADMIN.TOKEN },
-								});
-
-								await ws.subscribe({ collection: localCollectionCountries, uid: localCollectionCountries });
-								await ws.subscribe({ collection: localCollectionStates, uid: localCollectionStates });
-
-								const wsGql = createWebSocketGql(getUrl(vendor), {
-									auth: { access_token: USER.ADMIN.TOKEN },
-								});
-
-								await wsGql.subscribe({
-									collection: localCollectionCountries,
-									jsonQuery: {
-										id: true,
-										name: true,
-										event: true,
-									},
-									uid: localCollectionCountries,
-								});
-
-								await wsGql.subscribe({
-									collection: localCollectionStates,
-									jsonQuery: {
-										id: true,
-										name: true,
-										country_id: {
-											id: true,
-										},
-										event: true,
-									},
-									uid: localCollectionStates,
-								});
-
-								// Action
-								const response = await request(getUrl(vendor))
-									.post(`/items/${localCollectionStates}`)
-									.send(states)
-									.set('Authorization', `Bearer ${USER.ADMIN.TOKEN}`);
-
-								const mutationKey = `create_${localCollectionStates}_items`;
-
-								const gqlResponse = await requestGraphQL(getUrl(vendor), false, USER.ADMIN.TOKEN, {
-									mutation: {
-										[mutationKey]: {
-											__args: {
-												data: states2,
-											},
-											id: true,
-										},
-									},
-								});
-
-								ws.conn.close();
-								wsGql.client.dispose();
-
-								// Assert
-								expect(response.statusCode).toBe(400);
-								expect(response.body.errors).toBeDefined();
-
-								expect(response.body.errors[0].message).toBe(
-									`Invalid payload. Exceeded max batch mutation limit of ${config.envs[vendor]['MAX_BATCH_MUTATION']}.`,
-								);
-
-								expect(gqlResponse.statusCode).toBe(200);
-								expect(gqlResponse.body.errors).toBeDefined();
-
-								expect(gqlResponse.body.errors[0].message).toBe(
-									`Invalid payload. Exceeded max batch mutation limit of ${config.envs[vendor]['MAX_BATCH_MUTATION']}.`,
-								);
-
-								expect(ws.getMessageCount(localCollectionCountries)).toBe(1);
-								expect(ws.getMessageCount(localCollectionStates)).toBe(1);
-								expect(wsGql.getMessageCount(localCollectionCountries)).toBe(0);
-								expect(wsGql.getMessageCount(localCollectionStates)).toBe(0);
-							},
-							120000,
-						);
-					});
-				});
-
-				describe('updateBatch', () => {
-					describe('passes when below limit', () => {
-						it.each(vendors)(
-							'%s',
-							async (vendor) => {
-								// Setup
-								const count = Number(config.envs[vendor]['MAX_BATCH_MUTATION']) / 2;
-								const countCreate = Math.floor(count / 2);
-								const statesID = [];
-								const statesID2 = [];
-
-								for (let i = 0; i < count; i++) {
-									const state: any = createState(pkType);
-									state.name = `max_batch_mutation_${i.toString().padStart(3, '0')}`;
-
-									if (i >= countCreate) {
-										state.country_id = createCountry(pkType);
-									}
-
-									statesID.push((await CreateItem(vendor, { collection: localCollectionStates, item: state })).id);
-
-									const state2: any = createState(pkType);
-									state2.name = `max_batch_mutation_gql_${i.toString().padStart(3, '0')}`;
-
-									if (i >= countCreate) {
-										state2.country_id = createCountry(pkType);
-									}
-
-									statesID2.push((await CreateItem(vendor, { collection: localCollectionStates, item: state2 })).id);
-								}
-
-								const states = await ReadItem(vendor, {
-									collection: localCollectionStates,
-									fields: ['*', 'country_id.id', 'country_id.name'],
-									sort: ['name'],
-									filter: { id: { _in: statesID } },
-								});
-
-								const states2 = await ReadItem(vendor, {
-									collection: localCollectionStates,
-									fields: ['*', 'country_id.id', 'country_id.name'],
-									sort: ['name'],
-									filter: { id: { _in: statesID2 } },
-								});
-
-								for (let i = 0; i < states.length; i++) {
-									if (i < countCreate) {
-										states[i].country_id = createCountry(pkType);
-									} else {
-										states[i].country_id.name = 'updated';
-									}
-								}
-
-								for (let i = 0; i < states2.length; i++) {
-									if (i < countCreate) {
-										states2[i].country_id = createCountry(pkType);
-									} else {
-										states2[i].country_id.name = 'updated';
-									}
-								}
-
-								// Action
-								const response = await request(getUrl(vendor))
-									.patch(`/items/${localCollectionStates}`)
-									.send(states)
-									.set('Authorization', `Bearer ${USER.ADMIN.TOKEN}`);
-
-								const mutationKey = `update_${localCollectionStates}_batch`;
-
-								const gqlResponse = await requestGraphQL(getUrl(vendor), false, USER.ADMIN.TOKEN, {
-									mutation: {
-										[mutationKey]: {
-											__args: {
-												data: states2,
-											},
-											id: true,
-										},
-									},
-								});
-
-								// Assert
-								expect(response.statusCode).toBe(200);
-								expect(response.body.data.length).toBe(count);
-
-								expect(gqlResponse.statusCode).toBe(200);
-								expect(gqlResponse.body.data[mutationKey].length).toEqual(count);
-							},
-							120000,
-						);
-					});
-
-					describe('errors when above limit', () => {
-						it.each(vendors)(
-							'%s',
-							async (vendor) => {
-								// Setup
-								const count = Number(config.envs[vendor]['MAX_BATCH_MUTATION']) / 2 + 1;
-								const countCreate = Math.floor(count / 2);
-								const statesID = [];
-								const statesID2 = [];
-
-								for (let i = 0; i < count; i++) {
-									const state: any = createState(pkType);
-									state.name = `max_batch_mutation_${i.toString().padStart(3, '0')}`;
-
-									if (i >= countCreate) {
-										state.country_id = createCountry(pkType);
-									}
-
-									statesID.push((await CreateItem(vendor, { collection: localCollectionStates, item: state })).id);
-
-									const state2: any = createState(pkType);
-									state2.name = `max_batch_mutation_gql_${i.toString().padStart(3, '0')}`;
-
-									if (i >= countCreate) {
-										state2.country_id = createCountry(pkType);
-									}
-
-									statesID2.push((await CreateItem(vendor, { collection: localCollectionStates, item: state2 })).id);
-								}
-
-								const states = await ReadItem(vendor, {
-									collection: localCollectionStates,
-									fields: ['*', 'country_id.id', 'country_id.name'],
-									sort: ['name'],
-									filter: { id: { _in: statesID } },
-								});
-
-								const states2 = await ReadItem(vendor, {
-									collection: localCollectionStates,
-									fields: ['*', 'country_id.id', 'country_id.name'],
-									sort: ['name'],
-									filter: { id: { _in: statesID2 } },
-								});
-
-								for (let i = 0; i < states.length; i++) {
-									if (i < countCreate) {
-										states[i].country_id = createCountry(pkType);
-									} else {
-										states[i].country_id.name = 'updated';
-									}
-								}
-
-								for (let i = 0; i < states2.length; i++) {
-									if (i < countCreate) {
-										states2[i].country_id = createCountry(pkType);
-									} else {
-										states2[i].country_id.name = 'updated';
-									}
-								}
-
-								// Action
-								const response = await request(getUrl(vendor))
-									.patch(`/items/${localCollectionStates}`)
-									.send(states)
-									.set('Authorization', `Bearer ${USER.ADMIN.TOKEN}`);
-
-								const mutationKey = `update_${localCollectionStates}_batch`;
-
-								const gqlResponse = await requestGraphQL(getUrl(vendor), false, USER.ADMIN.TOKEN, {
-									mutation: {
-										[mutationKey]: {
-											__args: {
-												data: states2,
-											},
-											id: true,
-										},
-									},
-								});
-
-								// Assert
-								expect(response.statusCode).toBe(400);
-								expect(response.body.errors).toBeDefined();
-
-								expect(response.body.errors[0].message).toBe(
-									`Invalid payload. Exceeded max batch mutation limit of ${config.envs[vendor]['MAX_BATCH_MUTATION']}.`,
-								);
-
-								expect(gqlResponse.statusCode).toBe(200);
-								expect(gqlResponse.body.errors).toBeDefined();
-
-								expect(gqlResponse.body.errors[0].message).toBe(
-									`Invalid payload. Exceeded max batch mutation limit of ${config.envs[vendor]['MAX_BATCH_MUTATION']}.`,
-								);
-							},
-							120000,
-						);
-					});
-				});
-
-				describe('updateMany', () => {
-					describe('passes when below limit', () => {
-						it.each(vendors)(
-							'%s',
-							async (vendor) => {
-								// Setup
-								const count = Number(config.envs[vendor]['MAX_BATCH_MUTATION']) - 1;
-								const stateIDs = [];
-								const stateIDs2 = [];
-								const newCountry = createCountry(pkType);
-								const newCountry2 = createCountry(pkType);
-
-								for (let i = 0; i < count; i++) {
-									const state: any = createState(pkType);
-									state.country_id = createCountry(pkType);
-									stateIDs.push((await CreateItem(vendor, { collection: localCollectionStates, item: state })).id);
-
-									const state2: any = createState(pkType);
-									state2.country_id = createCountry(pkType);
-									stateIDs2.push((await CreateItem(vendor, { collection: localCollectionStates, item: state2 })).id);
-								}
-
-								const ws = createWebSocketConn(getUrl(vendor), {
-									auth: { access_token: USER.ADMIN.TOKEN },
-								});
-
-								await ws.subscribe({ collection: localCollectionCountries, uid: localCollectionCountries });
-								await ws.subscribe({ collection: localCollectionStates, uid: localCollectionStates });
-
-								const wsGql = createWebSocketGql(getUrl(vendor), {
-									auth: { access_token: USER.ADMIN.TOKEN },
-								});
-
-								const subscriptionKeyCountries = await wsGql.subscribe({
-									collection: localCollectionCountries,
-									jsonQuery: {
-										event: true,
-										data: {
-											id: true,
-											name: true,
-										},
-									},
-									uid: localCollectionCountries,
-								});
-
-								const subscriptionKeyStates = await wsGql.subscribe({
-									collection: localCollectionStates,
-									jsonQuery: {
-										event: true,
-										data: {
-											id: true,
-											name: true,
-											country_id: {
-												id: true,
-											},
-										},
-									},
-									uid: localCollectionStates,
-								});
-
-								// Action
-								const response = await request(getUrl(vendor))
-									.patch(`/items/${localCollectionStates}`)
-									.send({ keys: stateIDs, data: { country_id: newCountry } })
-									.set('Authorization', `Bearer ${USER.ADMIN.TOKEN}`);
-
-								const wsMessagesCountries = await ws.getMessages(1, { uid: localCollectionCountries });
-								const wsMessagesStates = await ws.getMessages(1, { uid: localCollectionStates });
-								const wsGqlMessagesCountries = await wsGql.getMessages(1, { uid: localCollectionCountries });
-								const wsGqlMessagesStates = await wsGql.getMessages(count - 1, { uid: localCollectionStates });
-
-								const mutationKey = `update_${localCollectionStates}_items`;
-
-								const gqlResponse = await requestGraphQL(getUrl(vendor), false, USER.ADMIN.TOKEN, {
-									mutation: {
-										[mutationKey]: {
-											__args: {
-												ids: stateIDs2,
-												data: { country_id: newCountry2 },
-											},
-											id: true,
-										},
-									},
-								});
-
-								const wsMessagesGqlCountries = await ws.getMessages(1, { uid: localCollectionCountries });
-								const wsMessagesGqlStates = await ws.getMessages(1, { uid: localCollectionStates });
-								ws.conn.close();
-								const wsGqlMessagesGqlCountries = await wsGql.getMessages(1, { uid: localCollectionCountries });
-								const wsGqlMessagesGqlStates = await wsGql.getMessages(count - 1, { uid: localCollectionStates });
-								wsGql.client.dispose();
-
-								// Assert
-								expect(response.statusCode).toBe(200);
-								expect(response.body.data.length).toBe(count);
-
-								expect(gqlResponse.statusCode).toBe(200);
-								expect(gqlResponse.body.data[mutationKey].length).toEqual(count);
-
-								for (const { messagesCountries, messagesStates } of [
-									{ messagesCountries: wsMessagesCountries, messagesStates: wsMessagesStates },
-									{ messagesCountries: wsMessagesGqlCountries, messagesStates: wsMessagesGqlStates },
-								]) {
-									expect(messagesCountries?.length).toBe(1);
-									expect(messagesStates?.length).toBe(1);
-
-									expect(messagesCountries![0]).toMatchObject({
-										type: 'subscription',
-										event: 'create',
-										data: [
-											{
-												id: expect.anything(),
-												name: expect.anything(),
-											},
-										],
-									});
-
-									expect(messagesStates![0]).toMatchObject({
-										type: 'subscription',
-										event: 'update',
-										data: expect.arrayContaining([
-											expect.objectContaining({
-												id: expect.anything(),
-												name: expect.any(String),
-												country_id: expect.anything(),
-											}),
-										]),
-									});
-								}
-
-								for (const { messagesCountries, messagesStates } of [
-									{ messagesCountries: wsGqlMessagesCountries, messagesStates: wsGqlMessagesStates },
-									{ messagesCountries: wsGqlMessagesGqlCountries, messagesStates: wsGqlMessagesGqlStates },
-								]) {
-									expect(messagesCountries?.length).toBe(1);
-									expect(messagesStates?.length).toBe(count - 1);
-
-									expect(messagesCountries![0]).toEqual({
-										data: {
-											[subscriptionKeyCountries]: {
-												event: 'create',
-												data: {
-													id: expect.anything(),
-													name: expect.any(String),
-												},
-											},
-										},
-									});
-
-									for (let i = 0; i < count - 1; i++) {
-										expect(messagesStates![i]).toEqual({
-											data: {
-												[subscriptionKeyStates]: {
-													event: 'update',
-													data: {
-														id: expect.anything(),
-														name: expect.any(String),
-														country_id: {
-															id: expect.anything(),
-														},
-													},
-												},
-											},
-										});
-									}
-								}
-							},
-							120_000,
-						);
-					});
-
-					describe('errors when above limit', () => {
-						it.each(vendors)(
-							'%s',
-							async (vendor) => {
-								// Setup
-								const count = Number(config.envs[vendor]['MAX_BATCH_MUTATION']);
-								const stateIDs = [];
-								const stateIDs2 = [];
-								const newCountry = createCountry(pkType);
-
-								for (let i = 0; i < count; i++) {
-									const state: any = createState(pkType);
-									state.country_id = createCountry(pkType);
-									stateIDs.push((await CreateItem(vendor, { collection: localCollectionStates, item: state })).id);
-
-									const state2: any = createState(pkType);
-									state2.country_id = createCountry(pkType);
-									stateIDs2.push((await CreateItem(vendor, { collection: localCollectionStates, item: state2 })).id);
-								}
-
-								const ws = createWebSocketConn(getUrl(vendor), {
-									auth: { access_token: USER.ADMIN.TOKEN },
-								});
-
-								await ws.subscribe({ collection: localCollectionCountries, uid: localCollectionCountries });
-								await ws.subscribe({ collection: localCollectionStates, uid: localCollectionStates });
-
-								const wsGql = createWebSocketGql(getUrl(vendor), {
-									auth: { access_token: USER.ADMIN.TOKEN },
-								});
-
-								await wsGql.subscribe({
-									collection: localCollectionCountries,
-									jsonQuery: {
-										event: true,
-										data: {
-											id: true,
-											name: true,
-										},
-									},
-									uid: localCollectionCountries,
-								});
-
-								await wsGql.subscribe({
-									collection: localCollectionStates,
-									jsonQuery: {
-										event: true,
-										data: {
-											id: true,
-											name: true,
-											country_id: {
-												id: true,
-											},
-										},
-									},
-									uid: localCollectionStates,
-								});
-
-								// Action
-								const response = await request(getUrl(vendor))
-									.patch(`/items/${localCollectionStates}`)
-									.send({ keys: stateIDs, data: { country_id: newCountry } })
-									.set('Authorization', `Bearer ${USER.ADMIN.TOKEN}`);
-
-								const mutationKey = `update_${localCollectionStates}_items`;
-
-								const gqlResponse = await requestGraphQL(getUrl(vendor), false, USER.ADMIN.TOKEN, {
-									mutation: {
-										[mutationKey]: {
-											__args: {
-												ids: stateIDs2,
-												data: { country_id: newCountry },
-											},
-											id: true,
-										},
-									},
-								});
-
-								ws.conn.close();
-								wsGql.client.dispose();
-
-								// Assert
-								expect(response.statusCode).toBe(400);
-								expect(response.body.errors).toBeDefined();
-
-								expect(response.body.errors[0].message).toBe(
-									`Invalid payload. Exceeded max batch mutation limit of ${config.envs[vendor]['MAX_BATCH_MUTATION']}.`,
-								);
-
-								expect(gqlResponse.statusCode).toBe(200);
-								expect(gqlResponse.body.errors).toBeDefined();
-
-								expect(gqlResponse.body.errors[0].message).toBe(
-									`Invalid payload. Exceeded max batch mutation limit of ${config.envs[vendor]['MAX_BATCH_MUTATION']}.`,
-								);
-
-								expect(ws.getMessageCount(localCollectionCountries)).toBe(1);
-								expect(ws.getMessageCount(localCollectionStates)).toBe(1);
-								expect(wsGql.getMessageCount(localCollectionCountries)).toBe(0);
-								expect(wsGql.getMessageCount(localCollectionStates)).toBe(0);
-							},
-							120000,
-						);
 					});
 				});
 			});
